requirements.txt
# Web Frameworks
flask==3.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
pymongo==4.6.0

# AI/LLM
google-generativeai==0.3.1
langchain==0.1.0

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1
python-barcode==0.15.1
Pillow==10.1.0

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development
pytest==7.4.3
httpx==0.25.2
